2to/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "2to"
version = "0.2.0"
description = """
Common utilities for 2_ converters (styles, caching, CLI helpers):
- [2pdf](../2pdf)
- [2html](../2html)
"""
readme = "README.md"
requires-python = ">=3.8"
authors = [{ name = "Henri J. Norden" }]
license = { text = "MIT" }
keywords = ["cli", "pandoc", "markdown", "styles", "css"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Environment :: Console",
]
dependencies = [
  "pypandoc-binary>=1.11",
]

[dependency-groups]
dev = ["aiohttp"]

[project.scripts]
2to-clear-cache = "from2to.cli_common:clear_cache_main"

[tool.setuptools.packages.find]
where = ["."]
include = ["from2to*"]

[tool.setuptools.package-data]
from2to = [
  "styles/included/*.css",
  "styles/**/*.md",
]
